Finalized collision manager and rebuilt nodes framework
works now more akin to what we discussed the other day.

A singular NodeManager owned by a player monitors and controls the nodes
that that player owns.

There is also a script graveyard for irrelevant scripts (depricated in
the past few days because functionality was absorbed into higher tier
scripts)